Сделал ограничение на размер загружаемых файлов

+++ pdf4/views.py
@@ -3,6 +3,7 @@ from .forms import MergeForm
 from .pdf4 import mergefunction, insertfunction, getnumpagesfunction, split1function, inimagesfunction, outimagesfunction, compressfunction, rotatefunction, inpdffunction, intextfunction
 from django.core.files.storage import FileSystemStorage
 
+
 def home_view(request):
 		return render(request, 'index.html')
 
@@ -12,6 +13,9 @@ def merge_view(request):
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']
         file2 = request.FILES['file2']
+        if (file1.size > 25*1024*1024) or (file2.size > 25*1024*1024):                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'merge.html')
         file1name = fs.save(file1.name, file1)  #сохраняем файлы из формы
         file2name = fs.save(file2.name, file2)
         resulturl = mergefunction('./pdf4/media/'+ file1name, './pdf4/media/'+ file2name)   #вызываем функцию для объединения файлов из pdf4.py, на выходе из которой получаем url результата
@@ -36,6 +40,9 @@ def insert_view(request):
             fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
             file1 = request.FILES['file1']
             file2 = request.FILES['file2']
+            if (file1.size > 25*1024*1024) or (file2.size > 25*1024*1024):                    #защита от отправки на сервер больших файлов
+                raise Exception('Слишком большой файл!')
+                return render(request, 'insert.html')
             file1name = fs.save(file1.name, file1)  #сохраняем файлы из формы
             file2name = fs.save(file2.name, file2)
             request.session['file1_cookie'] = file1name
@@ -68,6 +75,9 @@ def split_1_view(request):
     if request.method == 'POST' and request.FILES['file1']:
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']
+        if file1.size > 50*1024*1024:                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'split-1.html')
         file1name = fs.save(file1.name, file1)  #сохраняем файл из формы
         resulturl = split1function('./pdf4/media/'+ file1name)   #вызываем функцию для разделения файла на отдельные страницы из pdf4.py, на выходе из которой получаем url для скачивания архива со страницами
         fs.delete(file1name)    #удаляем исходный файлы
@@ -81,6 +91,9 @@ def in_images_view(request):
     if request.method == 'POST' and request.FILES['file1']:
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']  #передаем данные из формы в переменные
+        if file1.size > 50*1024*1024:                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'in-images.html')
         format = request.POST['format1']
         file1name = fs.save(file1.name, file1)  #сохраняем файл из переменной
         resulturl = inimagesfunction('./pdf4/media/'+ file1name, format)   #вызываем функцию для преобразования файла в картинки из pdf4.py, на выходе из которой получаем url для скачивания архива со страницами
@@ -96,18 +109,33 @@ def out_images_view(request):
         filename = dict.fromkeys([1,2,3,4,5]) #создем словарь для имен файлов, ключи словаря 1,2,3,4,5 а значения у ключей None
         if 'file1' in request.FILES:    #если вместо этого написать if request.FILES['file1']: то будет появляться ошибка в случае если поле пустое (если пользователь не выбрал файл)
             file1 = request.FILES['file1']
+            if file1.size > 25*1024*1024:                    #защита от отправки на сервер больших файлов
+                raise Exception('Слишком большой файл!')
+                return render(request, 'out-images.html')
             filename[1] = fs.save(file1.name, file1)  #сохраняем файлы из формы, имена файлов записываем в словарь
         if 'file2' in request.FILES:
             file2 = request.FILES['file2']
+            if file2.size > 25*1024*1024:                    #защита от отправки на сервер больших файлов
+                raise Exception('Слишком большой файл!')
+                return render(request, 'out-images.html')
             filename[2] = fs.save(file2.name, file2)  #сохраняем файлы из формы, имена файлов записываем в словарь
         if 'file3' in request.FILES:
             file3 = request.FILES['file3']
+            if file3.size > 25*1024*1024:                    #защита от отправки на сервер больших файлов
+                raise Exception('Слишком большой файл!')
+                return render(request, 'out-images.html')
             filename[3] = fs.save(file3.name, file3)  #сохраняем файлы из формы, имена файлов записываем в словарь
         if 'file4' in request.FILES:
             file4 = request.FILES['file4']
+            if file4.size > 25*1024*1024:                    #защита от отправки на сервер больших файлов
+                raise Exception('Слишком большой файл!')
+                return render(request, 'out-images.html')
             filename[4] = fs.save(file4.name, file4)  #сохраняем файлы из формы, имена файлов записываем в словарь
         if 'file5' in request.FILES:
             file5 = request.FILES['file5']
+            if file5.size > 25*1024*1024:                    #защита от отправки на сервер больших файлов
+                raise Exception('Слишком большой файл!')
+                return render(request, 'out-images.html')
             filename[5] = fs.save(file5.name, file5)  #сохраняем файлы из формы, имена файлов записываем в словарь
 
         resulturl = outimagesfunction(filename[1], filename[2], filename[3], filename[4], filename[5])   #вызываем функцию для сбора PDF-файла из картинок, на выходе из которой получаем url результата
@@ -134,6 +162,9 @@ def compress_view(request):
     if request.method == 'POST' and request.FILES['file1']:
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']  #передаем данные из формы в переменные
+        if file1.size > 50*1024*1024:                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'compress.html')
         level = request.POST['level1']
         file1name = fs.save(file1.name, file1)  #сохраняем файл из переменной
         resulturl = compressfunction('./pdf4/media/'+ file1name, level)   #вызываем функцию для сжатия файла из pdf4.py, на выходе из которой получаем url для скачивания результата
@@ -149,6 +180,9 @@ def rotate_view(request):
     if request.method == 'POST' and request.FILES['file1']:
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']  #передаем данные из формы в переменные
+        if file1.size > 50*1024*1024:                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'rotate.html')
         grad1 = request.POST['grad1']
         pages1 = request.POST['pages1']
         file1name = fs.save(file1.name, file1)  #сохраняем файл из переменной
@@ -163,6 +197,9 @@ def in_pdf_view(request):
     if request.method == 'POST' and request.FILES['file1']:
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']  #передаем данные из формы в переменные
+        if file1.size > 50*1024*1024:                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'in-pdf.html')
         file1name = fs.save(file1.name, file1)  #сохраняем файл из переменной
         resulturl = inpdffunction('./pdf4/media/'+ file1name)   #вызываем функцию для преобразования файла в PDF
         fs.delete(file1name)    #удаляем исходный файлы
@@ -175,6 +212,9 @@ def in_text_view(request):
     if request.method == 'POST' and request.FILES['file1']:
         fs = FileSystemStorage()        #создаем экземпляр джанго-класс для работы с файлами
         file1 = request.FILES['file1']  #передаем данные из формы в переменные
+        if file1.size > 50*1024*1024:                    #защита от отправки на сервер больших файлов
+            raise Exception('Слишком большой файл!')
+            return render(request, 'in-text.html')
         format = request.POST['format1']
         file1name = fs.save(file1.name, file1)  #сохраняем файл из переменной
         resulturl = intextfunction('./pdf4/media/'+ file1name, format)   #вызываем функцию для преобразования файла в текстовый формат из pdf4.py, на выходе из которой получаем url для скачивания текстового файла
